fix Coffee* and Ingredient* architecture

+++ src/main/java/band/effective/coffeeshop/model/dto/IngredientRequestDTO.java
@@ -6,14 +6,16 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 public class IngredientRequestDTO {
     private String name;
-    private Double quantity;
+    private BigDecimal quantity;
     @JsonAlias({"price_per_one", "cost_per_one"})
-    private Double costPerOne;
+    private BigDecimal costPerOne;
 
 }
